Split out getThemesLastPageForQuery
